Add `Time.Measure` functions leveraging system monotonic clocks.
This is a collection of functions relating to measuring intervals
of time elapsed during some portion of the program's execution.

Functions like `Time.Measure.nanoseconds -> (...)` can be used to
measure a synchronous operation. For an asynchronous operation,
functions like `Time.Measure.current_nanoseconds` can return a
value to be captured before and after the operation and
subtracted to indicate the elapsed time over that interval.