Add `Time.Duration` data type.
It uses the same internal numbers as the `Time` struct, but it has
the semantics of being a duration of time rather than an absolute
point in time relative to the start of the Common Era
in the proleptic Gregorian calendar (as `Time` does).

`Time.+` and `Time.-` have been changed to accept a `Time.Duration`
instead of an absolute `Time` (and they still return an absolute `Time`).